Add tests for interactions related to NewRelic tracker. Change signature of trackEvent method.

+++ RITracking/app/src/androidTest/java/de/rocketinternet/android/tracking/trackers/mocks/RIAdJustTrackerMock.java
@@ -44,7 +44,7 @@ public class RIAdJustTrackerMock extends RIAdjustTracker {
     }
 
     @Override
-    public void trackEvent(String event, int value, String action, String category, Map<String, Object> data) {
+    public void trackEvent(String event, long value, String action, String category, Map<String, Object> data) {
         mIsEventTracked = true;
         mNumberOfSentEvents++;
         mSignal.countDown();
