+++ RITracking/app/src/androidTest/java/de/rocketinternet/android/tracking/trackers/mocks/RIGoogleTagManagerTrackerMock.java
@@ -47,7 +47,7 @@ public class RIGoogleTagManagerTrackerMock extends RIGoogleTagManagerTracker {
     }
 
     @Override
-    public void trackEvent(String event, int value, String action, String category, Map<String, Object> data) {
+    public void trackEvent(String event, long value, String action, String category, Map<String, Object> data) {
         mIsEventTracked = true;
         mNumberOfSentEvents++;
         mSignal.countDown();
