+++ RITracking/app/src/main/java/de/rocketinternet/android/tracking/core/RITracking.java
@@ -36,10 +36,9 @@ import de.rocketinternet.android.tracking.utils.RIResourceUtils;
 
 /**
  * @author alessandro.balocco
- *         <p/>
- *         This class allows users of this library to interact with different tracking systems. The class is
- *         provides functionalities to track specific events and based on that it automatically spreads these
- *         events to registered tracking libraries.
+ * This class allows users of this library to interact with different tracking systems. The class is
+ * provides functionalities to track specific events and based on that it automatically spreads these
+ * events to registered tracking libraries.
  */
 public class RITracking implements
         RIEventTracking,
@@ -157,9 +156,10 @@ public class RITracking implements
     }
 
     @Override
-    public void trackEvent(final String event, final int value, final String action, final String category, final Map<String, Object> data) {
-        RILogUtils.logDebug("Tracking event: " + event + " with value: " + value + " with action: " + action +
-                "with category: " + category + " and data: " + data);
+    public void trackEvent(final String event, final long value, final String action,
+                           final String category, final Map<String, Object> data) {
+        RILogUtils.logDebug("Tracking event: " + event + " with value: " + value + " with action: "
+                + action + "with category: " + category + " and data: " + data);
 
         if (mTrackers == null) {
             RILogUtils.logError("Invalid call with non-existent trackers. Initialisation may have failed.");
@@ -345,7 +345,8 @@ public class RITracking implements
     }
 
     @Override
-    public void trackAddProductToCart(final RITrackingProduct product, final String cartId, final String location) {
+    public void trackAddProductToCart(final RITrackingProduct product, final String cartId,
+                                      final String location) {
         if (product != null) {
             RILogUtils.logDebug("Tracking add product with id " + product.getIdentifier() + " to cart");
 
@@ -368,7 +369,8 @@ public class RITracking implements
     }
 
     @Override
-    public void trackRemoveProductFromCart(final RITrackingProduct product, final int quantity, final double cartValue) {
+    public void trackRemoveProductFromCart(final RITrackingProduct product, final int quantity,
+                                           final double cartValue) {
         if (product != null) {
             RILogUtils.logDebug("Tracking remove " + quantity + " products with id " +
                     product.getIdentifier() + " from cart");
