+++ RITracking/app/src/main/java/de/rocketinternet/android/tracking/interfaces/RIEventTracking.java
@@ -11,7 +11,7 @@ public interface RIEventTracking {
 
     /**
      * A method to track an event happening inside the application.
-     * <p/>
+     *
      * The event may be triggered by the user and further information, such as category, action and
      * value are available.
      *
@@ -21,5 +21,5 @@ public interface RIEventTracking {
      * @param category (optional) An identifier for the category of the app the user is in
      * @param data     (optional) Additional data about the event
      */
-    void trackEvent(String event, int value, String action, String category, Map<String, Object> data);
+    void trackEvent(String event, long value, String action, String category, Map<String, Object> data);
 }
