+++ RITracking/app/src/main/java/de/rocketinternet/android/tracking/trackers/RIAd4PushTracker.java
@@ -123,7 +123,7 @@ public class RIAd4PushTracker extends RITracker implements
     }
 
     @Override
-    public void trackEvent(String event, int value, String action, String category, Map<String, Object> data) {
+    public void trackEvent(String event, long value, String action, String category, Map<String, Object> data) {
         RILogUtils.logDebug("Ad4Push tracker - Tracking event with name: " + event);
 
         if (mA4S == null) {
