+++ RITracking/app/src/main/java/de/rocketinternet/android/tracking/trackers/RIGoogleAnalyticsTracker.java
@@ -73,7 +73,7 @@ public class RIGoogleAnalyticsTracker extends RITracker implements
     }
 
     @Override
-    public void trackEvent(String event, int value, String action, String category, Map<String, Object> data) {
+    public void trackEvent(String event, long value, String action, String category, Map<String, Object> data) {
         RILogUtils.logDebug("Google Analytics - Tracking event: " + event);
 
         if (mTracker == null) {
